+++ apps/web/types/categories.ts
@@ -1,0 +1,6 @@
+export interface categoriesDataSnake {
+  id: number;
+  name: string;
+  sort_order: number;
+  visibility: string;
+}
